For QAPAAS-36 and SVY-11826
When trying to trigger a parameterized build without providing a valid
Jenkins-Crumb header, Jenkins would give an error code but no hint to
the reason in the body; but it did say crumb is needed in the
statusReasonPhrase.

So as that is helpful information I added a way to get it.