SVY-16030 Enable TLS on Rabbit MQ
added 2 more properties:

amqpbroadcaster.keystore.path
amqpbroadcaster.keystore.password

if those are set then the SSLContext is initialized with that keystore
for a default or given TLS protocol.

If only the amqpbroadcaster.tlsprotocols is given and noth that keystore
path/password then this willl enable TLS in a Trust ALL mode. So this
should only be used in internal networks